add manage facilites tab
